# Conflicts:
#	src/main/java/com/forgeessentials/commands/item/CommandCraft.java
#	src/main/java/com/forgeessentials/commands/item/CommandDrop.java
#	src/main/java/com/forgeessentials/commands/item/CommandRename.java
#	src/main/java/com/forgeessentials/commands/item/CommandRepair.java
#	src/main/java/com/forgeessentials/commands/item/CommandVirtualchest.java
#	src/main/java/com/forgeessentials/commands/player/CommandAFK.java
#	src/main/java/com/forgeessentials/commands/player/CommandBubble.java
#	src/main/java/com/forgeessentials/commands/player/CommandBurn.java
#	src/main/java/com/forgeessentials/commands/player/CommandCapabilities.java
#	src/main/java/com/forgeessentials/commands/player/CommandDoAs.java
#	src/main/java/com/forgeessentials/commands/player/CommandHeal.java
#	src/main/java/com/forgeessentials/commands/player/CommandInventorySee.java
#	src/main/java/com/forgeessentials/commands/player/CommandKill.java
#	src/main/java/com/forgeessentials/commands/player/CommandLocate.java
#	src/main/java/com/forgeessentials/commands/player/CommandNoClip.java
#	src/main/java/com/forgeessentials/commands/player/CommandPotion.java
#	src/main/java/com/forgeessentials/commands/player/CommandSeen.java
#	src/main/java/com/forgeessentials/commands/player/CommandSmite.java
#	src/main/java/com/forgeessentials/commands/player/CommandSpeed.java
#	src/main/java/com/forgeessentials/commands/player/CommandVanish.java
#	src/main/java/com/forgeessentials/commands/server/CommandDelayedAction.java
#	src/main/java/com/forgeessentials/commands/server/CommandModlist.java
#	src/main/java/com/forgeessentials/commands/server/CommandPing.java
#	src/main/java/com/forgeessentials/commands/server/CommandRules.java
#	src/main/java/com/forgeessentials/commands/server/CommandServerSettings.java
#	src/main/java/com/forgeessentials/commands/world/CommandButcher.java
#	src/main/java/com/forgeessentials/commands/world/CommandFindblock.java
#	src/main/java/com/forgeessentials/core/preloader/mixin/block/MixinBlockEndPortal.java
#	src/main/java/com/forgeessentials/playerlogger/command/CommandPlayerlogger.java
#	src/main/java/com/forgeessentials/protection/ProtectionEventHandler.java
#	src/main/java/com/forgeessentials/scripting/ScriptEventHandler.java
#	src/main/java/com/forgeessentials/teleport/CommandBed.java
#	src/main/java/com/forgeessentials/teleport/CommandHome.java
#	src/main/java/com/forgeessentials/teleport/CommandSetSpawn.java
#	src/main/java/com/forgeessentials/teleport/CommandSpawn.java
#	src/main/java/com/forgeessentials/teleport/CommandTop.java
#	src/main/java/com/forgeessentials/teleport/CommandTp.java
#	src/main/java/com/forgeessentials/teleport/CommandTppos.java
#	src/main/java/com/forgeessentials/teleport/portal/PortalManager.java

+++ src/main/java/com/forgeessentials/core/misc/TeleportHelper.java
@@ -21,11 +21,11 @@ import net.minecraft.world.Teleporter;
 import net.minecraft.world.WorldProvider;
 import net.minecraft.world.WorldServer;
 import net.minecraftforge.common.DimensionManager;
+import net.minecraftforge.fe.event.entity.EntityPortalEvent;
 import net.minecraftforge.fml.common.FMLCommonHandler;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
-import net.minecraftforge.fe.event.entity.EntityPortalEvent;
 
 import com.forgeessentials.api.APIRegistry;
 import com.forgeessentials.api.UserIdent;
@@ -54,7 +54,7 @@ public class TeleportHelper extends ServerEventHandler
             int i = MathHelper.floor_double(entity.posX);
             int j = MathHelper.floor_double(entity.posY) - 1;
             int k = MathHelper.floor_double(entity.posZ);
-            entity.setLocationAndAngles((double)i, (double)j, (double)k, entity.rotationYaw, 0.0F);
+            entity.setLocationAndAngles(i, j, k, entity.rotationYaw, 0.0F);
         }
 
         @Override
@@ -143,7 +143,8 @@ public class TeleportHelper extends ServerEventHandler
             throw new TranslatedCommandException("You are not allowed to teleport from here.");
         if (!APIRegistry.perms.checkUserPermission(ident, point.toWorldPoint(), TELEPORT_TO))
             throw new TranslatedCommandException("You are not allowed to teleport to that location.");
-        if (player.dimension != point.getDimension()) {
+        if (player.dimension != point.getDimension())
+        {
             if (!APIRegistry.perms.checkPermission(player, TELEPORT_CROSSDIM_FROM))
                 throw new TranslatedCommandException("You are not allowed to teleport from this dimension.");
             if (!APIRegistry.perms.checkUserPermission(ident, point.toWorldPoint(), TELEPORT_CROSSDIM_TO))
@@ -270,7 +271,8 @@ public class TeleportHelper extends ServerEventHandler
             e.setCanceled(true);
         if (!APIRegistry.perms.checkUserPermission(ident, pointTo, TELEPORT_PORTALTO))
             e.setCanceled(true);
-        if (e.world.provider.getDimensionId() != e.targetDimension) {
+        if (e.world.provider.getDimensionId() != e.targetDimension)
+        {
             if (!APIRegistry.perms.checkUserPermission(ident, pointFrom, TELEPORT_CROSSDIM_PORTALFROM))
                 e.setCanceled(true);
             if (!APIRegistry.perms.checkUserPermission(ident, pointTo, TELEPORT_CROSSDIM_PORTALTO))
